Clean up build process for es/non-es builds
