+++ .gitignore
@@ -1,5 +1,4 @@
 node_modules
-/module
 /dist
 /docs
 /coverage
