+++ jsdoc.conf.json
@@ -13,31 +13,31 @@
         "plugins/markdown"
     ],
     "templates": {
-        "cleverLinks"   : false,
+        "cleverLinks": false,
         "monospaceLinks": false,
-        "default"       : {
-            "outputSourceFiles" : false
+        "default": {
+            "outputSourceFiles": false
         },
         "applicationName": "PIXI.sound",
-        "footer"        : "Made with ♥ by Matt Karl (github.com/bigtimebuddy)",
-        "copyright"     : "PIXI Copyright &copy; 2017 Matt Karl.",
-        "linenums"      : false,
+        "footer": "Made with ♥ by Matt Karl (github.com/bigtimebuddy)",
+        "copyright": "PIXI Copyright &copy; 2017 Matt Karl.",
+        "linenums": false,
         "meta": {
             "title": "PIXI.sound API Documentation",
             "description": "Documentation for PIXI.sound library",
             "keyword": "docs, documentation, pixi, pixijs, webaudio, html5, audio, javascript, jsdoc"
         }
     },
-    "markdown"  : {
-        "parser"        : "gfm",
-        "hardwrap"      : false
+    "markdown": {
+        "parser": "gfm",
+        "hardwrap": false
     },
     "opts": {
-        "encoding"      : "utf8",
-        "recurse"       : true,
-        "private"       : false,
-        "lenient"       : true,
-        "destination"   : "./docs",
-        "template"      : "./node_modules/jaguarjs-jsdoc"
+        "encoding": "utf8",
+        "recurse": true,
+        "private": false,
+        "lenient": true,
+        "destination": "./docs",
+        "template": "./node_modules/jaguarjs-jsdoc"
     }
 }
