+++ package.json
@@ -2,26 +2,26 @@
   "name": "pixi-sound",
   "version": "1.4.2",
   "description": "WebAudio API playback library with filters",
-  "main": "module/index.js",
-  "module": "module/index.es.js",
-  "types": "module/types/index.d.ts",
+  "main": "dist/pixi-sound.min.js",
+  "module": "dist/pixi-sound.es.min.js",
+  "types": "dist/types/index.d.ts",
   "scripts": {
     "pretest": "npm run build",
     "test": "npm run test:module && npm run test:module:legacy",
     "test:module": "floss --path=test/module.js",
     "test:module:legacy": "floss --path=test/module.legacy.js",
-    "coverage": "floss --path=test/coverage.js -c \"dist/pixi-sound.cjs.js\" -s",
-    "precoverage:report": "rimraf coverage",
-    "coverage:report": "npm run coverage -- -h",
-    "clean": "rimraf dist/** module/**",
+    "precoverage": "npm run build:dev -- -o coverage/pixi-sound.js -f cjs -e es6-object-assign,promise-polyfill",
+    "coverage": "floss --path=test/coverage.js -c \"coverage/pixi-sound.js\" -s",
+    "clean": "rimraf dist/**",
     "start": "npm run watch",
     "watch": "npm run build:dev -- --watch",
-    "lint": "tslint examples/**.js test/*.js src/**.ts --format=stylish",
+    "lint": "tslint examples/**.js test/*.js src/**.ts --format=stylish --fix",
     "prebuild": "npm run clean && npm run lint",
-    "build": "npm run build:dev && npm run build:prod && npm run build:module",
+    "build": "npm run build:dev && npm run build:prod && npm run build:es:dev && npm run build:es:prod",
     "build:dev": "rollup -c --environment BUILD:dev",
     "build:prod": "rollup -c --environment BUILD:prod",
-    "build:module": "rollup -c rollup.config.module.js",
+    "build:es:dev": "npm run build:dev -- -f es -o dist/pixi-sound.es.js",
+    "build:es:prod": "npm run build:prod -- -f es -o dist/pixi-sound.es.min.js",
     "predocs": "rimraf docs/** .docs && gulp comments",
     "docs": "jsdoc -c jsdoc.conf.json -R README.md",
     "postdocs": "rimraf .docs",
@@ -41,7 +41,6 @@
   },
   "files": [
     "dist/",
-    "module/",
     "LICENSE",
     "package.json",
     "README.md"
@@ -62,6 +61,7 @@
     "@types/pixi.js": "^4.5.1",
     "@types/es6-shim": "^0.31.34",
     "electron": "^1.4.15",
+    "es6-object-assign": "^1.1.0",
     "floss": "^2.0.1",
     "gh-pages": "^0.12.0",
     "gulp": "^3.9.1",
@@ -69,6 +69,7 @@
     "jaguarjs-jsdoc": "^1.0.2",
     "jsdoc": "^3.4.3",
     "pixi.js": "^4.3.4",
+    "promise-polyfill": "^6.0.2",
     "rimraf": "^2.5.4",
     "rollup": "^0.43.0",
     "rollup-plugin-commonjs": "^8.0.2",
@@ -81,9 +82,5 @@
   },
   "peerDependencies": {
     "pixi.js": "^4.3.4"
-  },
-  "dependencies": {
-    "es6-object-assign": "^1.1.0",
-    "promise-polyfill": "^6.0.2"
   }
 }
