+++ test/legacy.js
@@ -1,0 +1,1 @@
+require("./")("../", true);
