--- test/module.legacy.js
@@ -1,1 +1,0 @@
-require("./")("../module", true);
