--- tsconfig.module.json
@@ -1,22 +1,0 @@
-{
-    "compilerOptions": {
-        "moduleResolution": "node",
-        "target": "es5",
-        "removeComments": true,
-        "noImplicitAny": true,
-        "preserveConstEnums": true,
-        "sourceMap": true,
-        "inlineSources": true,
-        "declaration": true,
-        "declarationDir": "module/types"
-    },
-    "files": [
-        "typings/index.d.ts"
-    ],
-    "include": [
-        "src/**/*"
-    ],
-    "exclude": [
-        "node_modules"
-    ]
-}
